Update RelationshipCanvas.java
Relationship panel fixed
